ai-service/requirements.txt
flask==3.0.2
flask-cors==4.0.0
PyPDF2==3.0.1
spacy==3.7.4
python-dotenv==1.0.1
gunicorn==21.2.0
